# hospedate-vscode/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/reviews/{id}:
    get:
      tags:
      - Reviews
      summary: Get review by ID
      operationId: getReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
    put:
      tags:
      - Reviews
      summary: Update a review
      operationId: updateReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "401":
          description: Unauthorized – login required
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "403":
          description: Forbidden – not the author
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
    delete:
      tags:
      - Reviews
      summary: Delete a review
      operationId: deleteReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized – login required
        "403":
          description: Forbidden – not the author
        "404":
          description: Not found
...
  /api/v1/hotels/{id}:
    get:
      tags:
      - Hotels
      summary: Get hotel by ID
      operationId: getHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HotelDTO"
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HotelDTO"
    put:
      tags:
      - Hotels
      summary: Update a hotel
      operationId: updateHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HotelDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HotelDTO"
        "400":
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HotelDTO"
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HotelDTO"
    delete:
      tags:
      - Hotels
      summary: Delete a hotel
      operationId: deleteHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /api/v1/admin/users/{id}:
    get:
      tags:
      - Admin
      summary: Get user by ID
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
        "403":
          description: Forbidden – admin only
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    put:
      tags:
      - Admin
      summary: Update a user
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
        "409":
          description: Email conflict
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - Admin
      summary: Delete a user
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Deleted
        "403":
          description: Cannot delete admin
        "404":
          description: Not found
  /api/v1/admin/reserves/{id}:
    get:
      tags:
      - Admin
      summary: Get reserve by ID
      operationId: getReserve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReserveDTO"
        "403":
          description: Forbidden – admin only
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReserveDTO"
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReserveDTO"
    put:
      tags:
      - Admin
      summary: Update a reserve
      operationId: updateReserve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - Admin
      summary: Delete a reserve
      operationId: deleteReserve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /api/v1/admin/users:
    get:
      tags:
      - Admin
      summary: List all non-admin users (paginated)
      operationId: listUsers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserDTO"
        "403":
          description: Forbidden – admin only
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserDTO"
  /api/v1/admin/reserves:
    get:
      tags:
      - Admin
      summary: List all reserves (paginated)
      operationId: listReserves
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageReserveDTO"
        "403":
          description: Forbidden – admin only
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageReserveDTO"
  /api/v1/reserves/{id}:
    delete:
      tags:
      - Reserves
      summary: Delete a reserve
      operationId: deleteReserve_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden – not the owner
        "404":
          description: Not found
components:
  schemas:
    ReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          maxLength: 100
          minLength: 0
        comment:
          type: string
          maxLength: 2000
          minLength: 0
        rating:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
        publishDate:
          type: string
          format: date
        authorName:
          type: string
        hotelId:
          type: integer
          format: int64
      required:
      - comment
      - hotelId
      - title
    ReserveDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        entryDate:
          type: string
          format: date
        departureDate:
          type: string
          format: date
        guests:
          type: integer
          format: int32
          minimum: 1
        nights:
          type: integer
          format: int64
        price:
          type: number
          format: double
        status:
          type: string
        hotelId:
          type: integer
          format: int64
        hotelName:
          type: string
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
      required:
      - departureDate
      - entryDate
      - hotelId
    HotelDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
          minLength: 0
        tipo:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        price:
          type: number
          format: double
          minimum: 0
        description:
          type: string
          maxLength: 2000
          minLength: 0
        rating:
          type: number
          format: double
          maximum: 5
          minimum: 0
        wifi:
          type: boolean
        tv:
          type: boolean
        airConditioning:
          type: boolean
        family:
          type: boolean
        services:
          type: array
          items:
            type: string
          uniqueItems: true
        mainImageUrl:
          type: string
      required:
      - city
      - description
      - location
      - name
      - tipo
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
          minLength: 0
        email:
          type: string
          format: email
          minLength: 1
        role:
          type: string
      required:
      - email
      - name
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageReserveDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReserveDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
...
    PageUserDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean